Additional changes added

+++ quake/forward/topography.h
@@ -83,7 +83,7 @@ int    topo_setrec              ( octant_t *leaf, double    ticksize, edata_t  *
 double point_elevation          ( double xo, double yo );
 int    topo_nodesearch          ( tick_t x, tick_t y, tick_t z, double ticksize );
 int    topo_crossings           ( double xo, double yo, double zo, double esize );
-int    topo_maxLevel();
+int8_t    topo_maxLevel();
 
 /* ERASE LATER THESE FUNCTIONS, ARE JUST FOR A QUICK CHECKING*/
 void topo_DRM_init ( mesh_t *myMesh, mysolver_t *mySolver);
